requirements.txt
discord.py
python-chess
svglib 
python-dotenv
reportlab